Working Through Fundamental Steps
Got a list of the stuff required to do the minimum (create template, create member of template, export member).
Going through it in order to create the functionality for the minimum viable.
A lot of it requiring making shared functionality which will speed up making the non-essential parts work anyways.