Template & Member Windows Working
In terms of data-editing at least. Validation rules not implemented yet due to needing to figure out how to best handle that. But, though the indices get a bit janky due to the template selection dropdown auto-alphabetising and the selection being based on index and not an ID, it works for data-entry.