Enter Parameters Of Validation Rules
Went for an awful hard-coded version for the sake of getting this over the finish line. Not memory-efficient, as a buffer is created for every validation rule which has input data, even if they're a different type and will never be active.
Also the interface layout is ugly as sin, so that needs to be tidied with more custom layout implementation (possible, just not done yet).