Validation Rules Implemented & Clearing Buffers
Did it the hard-coded way because the fancier cursed way I was trying to do it before wasn't working (aka couldn't figure out the cause of a bad cast error and don't have time to investigate it more thoroughly).
Also realised that buffers weren't being properly cleared when backspacing, so I have now implemented a manual clearing for that.