Enhanced Group Functionality
Removing items from a group is the majority of the efforts here, but also adding it (in the most horrifically ugly way) has been added.
Couple of bugs from the fact that adding new things requires an early return/refresh of the window (ie that mouse clicking doesn't work until you leave the window and come back to it), but that's markable as a known issue.