Added Comments & Fixed Bug With ActiveWindows
Windows weren't being properly removed from the active windows vector when they were closed. Fixed that because it's been bugging me for weeks.
The windows change back to default/starting position sometimes when a new window is opened or an old one is closed. I am not gonna have time to fix that in time for testing, 'cause I don't know where to start with it.